refactor - move all views into one python file
